Move chain interaction loop to new file `chain_leader.rs`
