+++ linera-service/src/lib.rs
@@ -2,6 +2,7 @@
 // Copyright (c) Zefchain Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod chain_leader;
 pub mod config;
 pub mod grpc_proxy;
 pub mod node_service;
